# Fix backend perf of report table