# Description

After a few hours of investigation with help from @jshoe , I concluded that 
1) `report_info` endpoint was the critical endpoint for the report table, taking up to 16s 
2) `fill_lineage_details` method was wasting the most time in `report_info`

I made a couple of changes to `fill_lineage_details` method to make `report_info` 2x faster (in local dev). 

1) avoiding active record and `as_json` by using `pluck` (6s)
2) using `Set` instead of array includes in inner loop (2s)

NOTE: 8s is still pretty slow for a webpage. For next steps, I recommend 

1) use gem OJ for faster json serialization (should get us 30% faster serialization across the whole app, see https://medium.com/infosimples/performance-comparison-for-json-generation-in-ruby-cc2cce55cf0b)
2) pre-compute `fill_lineage_details` on every pipeline run success
3) use gem 'fast_jsonapi' for even faster serialization of schematized json responses

# Test and Reproduce
Outline the steps to test or reproduce the PR here.

1. try `report_info` endpoint for a very large sample such as id 10399 (14270 total rows)
2. see that TTL is reduced by 50% from ~16s to ~8s
3. see same report table